Changed line color to light gray as it looks cleaner, fixed blurry svgs in chrome

+++ Source/AlphaTab/Platform/Svg/SvgCanvas.cs
@@ -60,7 +60,7 @@ namespace AlphaTab.Platform.Svg
             Buffer.Append("<svg xmlns=\"http://www.w3.org/2000/svg\" version=\"1.1\" width=\"");
             Buffer.Append(width);
             Buffer.Append("px\" height=\"");
-            Buffer.Append(height);
+            Buffer.Append((int)height);
             Buffer.Append("px\" class=\"alphaTabSurfaceSvg\">\n");
             _currentPath = new StringBuilder();
             _currentPathIsEmpty = true;
